Fix examples, update requirejs i18n to work
The 2.0.1 version of requirejs's i18n plugin did not work. I'm guessing
the plugin might actually depend on requirejs itself (I can find no
other reason for it not working standalone), but version 1.0.5 works
fine. As far as I can tell, the two versions are not that different, so
this is no big deal.

The inbrowser example was updated to work using the 0.8.3 coweb-jsoe
build.